Add database initialise to seeds

+++ db/seeds.rb
@@ -1,7 +1,10 @@
 puts '======> Destroying existing records'
 
 ActiveRecord::Base.connection.tables.each do |table|
-  ActiveRecord::Base.connection.execute("DELETE from #{table}")
+  unless table = 'ar_internal_metadata' || table = 'schema_migrations'
+    ActiveRecord::Base.connection.execute("DELETE from #{table}")
+    ActiveRecord::Base.connection.reset_pk_sequence! table
+  end
 end
 
 puts '======> Seeding books'
